+++ node/mcp-server.rest-app.posts/HISTORY.md
@@ -2,8 +2,8 @@
 
 ## HISTORY
 
-1. Copy+Pasted [modelcontextprotocol-typescript-sdk-example](../modelcontextprotocol-typescript-sdk-example)
+1. Copy+Pasted [mcp-server.examples.quick-start](../mcp-server.examples.quick-start)
 
 ### Changes
 
-1. TBD
+1. Added `GET`, `POST` `PUT` and `DELETE` methods for managing blog posts.